refactor: moved config lint to command group
