+++ src/commands/dev/mod.rs
@@ -1,10 +1,8 @@
 mod config;
-mod config_lint;
+mod lint;
 
 use crate::error::CliError;
 use clap::{Args, Subcommand};
-use config_lint::handle_config_lint;
-use std::path::PathBuf;
 
 #[derive(Debug, Args)]
 pub struct Dev {
@@ -15,11 +13,7 @@ pub struct Dev {
 #[derive(Debug, Subcommand)]
 pub enum DevSubcommand {
     /// Linting the config and show possible warnings, as well as suggestion how to fix the config file.
-    ConfigLint {
-        /// The path to the project
-        #[arg(default_value = ".")]
-        path: PathBuf,
-    },
+    Lint(lint::Lint),
     /// This command group contains the commands to interact with the config secrets with bunker.
     Config(config::Config),
 }
@@ -27,7 +21,7 @@ pub enum DevSubcommand {
 impl Dev {
     pub async fn handle_command(&self) -> Result<bool, CliError> {
         match &self.subcommand {
-            DevSubcommand::ConfigLint { path } => handle_config_lint(path),
+            DevSubcommand::Lint(config) => config.handle_command().await,
             DevSubcommand::Config(config) => config.handle_command().await,
         }
     }
